requirement.txt
sentence-transformers-5.0.0
langchain-community-0.3.26
datasets-3.3.2
transformers-4.49.0
peft-0.15.2
torch-2.5.1
gradio-5.35.0
nltk-3.9.1
rouge-score-0.1.2